+++ src/ResourceGraph/ResourceGraph/ChangeLog.md
@@ -18,6 +18,8 @@
         - Additional information about change #1
 -->
 ## Upcoming Release
+
+## Version 0.7.3
 * Improving logic of getting subscriptions for query to ARG
 
 ## Version 0.7.2